requirements.txt
# FGPS Core Dependencies
formalgeo>=0.0.4
psutil>=5.8.0
func-timeout>=4.3.5
sympy>=1.10

# AI Model Dependencies (Optional - for gemini/ directory)
google-generativeai>=0.3.0
openai>=1.0.0
pillow>=9.0.0
tqdm>=4.64.0
pydantic>=2.0.0

# Additional utilities
numpy>=1.21.0

